@@ -1,19 +1,19 @@
 /*!
  * Copyright (c) 2020-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const {agent} = require('bedrock-https-agent');
-const bedrock = require('bedrock');
+import * as bedrock from '@bedrock/core';
+import * as helpers from './helpers.js';
+import {agent} from '@bedrock/https-agent';
+import {createRequire} from 'module';
+import {documentLoader as brDocLoader} from '@bedrock/jsonld-document-loader';
+import {mockData} from './mock.data.js';
+const require = createRequire(import.meta.url);
 const {CapabilityAgent} = require('@digitalbazaar/webkms-client');
 const didKeyDriver = require('@digitalbazaar/did-method-key').driver();
-const {documentLoader: brDocLoader} =
-  require('bedrock-jsonld-document-loader');
 const {Ed25519Signature2020} = require('@digitalbazaar/ed25519-signature-2020');
-const helpers = require('./helpers');
 const {httpClient} = require('@digitalbazaar/http-client');
 const vc = require('@digitalbazaar/vc');
-const mockData = require('./mock.data');
+
 const {util: {clone}} = bedrock;
 
 const {baseUrl} = mockData;
