@@ -1,10 +1,12 @@
 /*!
  * Copyright (c) 2022 Digital Bazaar, Inc. All rights reserved.
  */
+import * as bedrock from '@bedrock/core';
+import * as database from '@bedrock/mongodb';
 import assert from 'assert-plus';
-import bedrock from 'bedrock';
-import database from 'bedrock-mongodb';
-import {generateId, decodeId} from 'bnid';
+import {createRequire} from 'module';
+const require = createRequire(import.meta.url);
+const {generateId, decodeId} = require('bnid');
 
 const {util: {BedrockError}} = bedrock;
 
