@@ -1,9 +1,8 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import bedrock from 'bedrock';
-
-const {config} = bedrock;
+import {config} from '@bedrock/core';
+import '@bedrock/app-identity';
 
 const cfg = config['vc-verifier'] = {};
 
