@@ -1,21 +1,21 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import 'bedrock-credentials-context';
-import 'bedrock-did-context';
-import 'bedrock-did-io';
-import 'bedrock-veres-one-context';
-import 'bedrock-vc-status-list-context';
-import 'bedrock-vc-revocation-list-context';
-import 'bedrock-security-context';
-import bedrock from 'bedrock';
-import {createContextDocumentLoader} from 'bedrock-service-context-store';
-import {didIo} from 'bedrock-did-io';
+import * as bedrock from '@bedrock/core';
+import '@bedrock/credentials-context';
+import '@bedrock/did-context';
+import '@bedrock/did-io';
+import '@bedrock/security-context';
+import '@bedrock/vc-status-list-context';
+import '@bedrock/vc-revocation-list-context';
+import '@bedrock/veres-one-context';
+import {createContextDocumentLoader} from '@bedrock/service-context-store';
+import {didIo} from '@bedrock/did-io';
 import {
   documentLoader as brDocLoader,
   httpClientHandler,
   JsonLdDocumentLoader
-} from 'bedrock-jsonld-document-loader';
+} from '@bedrock/jsonld-document-loader';
 
 const serviceType = 'vc-verifier';
 let webLoader;
