@@ -1,8 +1,8 @@
 /*!
  * Copyright (c) 2018-2022 Digital Bazaar, Inc. All rights reserved.
  */
+import * as bedrock from '@bedrock/core';
 import {createChallenge, verifyChallenge} from './challenges.js';
-import bedrock from 'bedrock';
 import {asyncHandler} from 'bedrock-express';
 import bodyParser from 'body-parser';
 import {checkStatus} from './status.js';
@@ -13,11 +13,15 @@ import {
   verifyPresentationBody
 } from '../schemas/bedrock-vc-verifier.js';
 import {createDocumentLoader} from './documentLoader.js';
-import {createValidateMiddleware as validate} from 'bedrock-validation';
-import {Ed25519Signature2020} from '@digitalbazaar/ed25519-signature-2020';
-import {Ed25519Signature2018} from '@digitalbazaar/ed25519-signature-2018';
-import {metering, middleware} from 'bedrock-service-core';
-import vc from '@digitalbazaar/vc';
+import {createRequire} from 'module';
+import {createValidateMiddleware as validate} from '@bedrock/validation';
+import {metering, middleware} from '@bedrock/service-core';
+const require = createRequire(import.meta.url);
+const {Ed25519Signature2020} =
+  require('@digitalbazaar/ed25519-signature-2020');
+const {Ed25519Signature2018} =
+  require('@digitalbazaar/ed25519-signature-2018');
+const vc = require('@digitalbazaar/vc');
 
 const {util: {BedrockError}} = bedrock;
 
