@@ -2,14 +2,16 @@
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
 import assert from 'assert-plus';
-import {
-  checkStatus as statusListCheckStatus,
-  statusTypeMatches as statusListStatusTypeMatches
-} from '@digitalbazaar/vc-status-list';
-import {
-  checkStatus as revocationListCheckStatus,
-  statusTypeMatches as revocationListStatusTypeMatches
-} from 'vc-revocation-list';
+import {createRequire} from 'module';
+const require = createRequire(import.meta.url);
+const {
+  checkStatus: statusListCheckStatus,
+  statusTypeMatches: statusListStatusTypeMatches
+} = require('@digitalbazaar/vc-status-list');
+const {
+  checkStatus: revocationListCheckStatus,
+  statusTypeMatches: revocationListStatusTypeMatches
+} = require('vc-revocation-list');
 
 const handlerMap = new Map();
 handlerMap.set('RevocationList2020Status', {
